requirements.txt
asgiref==3.7.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dash==0.29.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
django-filter==24.2
firebase-admin==6.5.0
openapi-codec==1.3.2
packaging==21.3
Pillow==10.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
pandas==2.2.2
requests==2.25.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.6.0